+++ src/django/api/management/commands/loadfixtures.py
@@ -17,6 +17,7 @@ class Command(BaseCommand):
                          'users.json',
                          'contributors.json',
                          'facility_lists.json',
+                         'sources.json',
                          'facility_list_items.json')
             if match:
                 call_command('loaddata',