+++ src/django/api/reports/average_list_count_per_contributor.sql
@@ -7,10 +7,11 @@ FROM (
     c.id,
     COUNT(*) AS list_count
   FROM api_facilitylist l
-  JOIN api_contributor c ON c.id = l.contributor_id
+  JOIN api_source s ON s.facility_list_id = l.id
+  JOIN api_contributor c ON c.id = s.contributor_id
   WHERE date_part('month', l.created_at) != date_part('month', now())
-  AND l.is_public = 't'
-  AND l.is_active = 't'
+  AND s.is_public = 't'
+  AND s.is_active = 't'
   GROUP BY date_part('month', l.created_at), c.id
 ) s
 GROUP BY month